Update sketch.js
We added a dual mode featuring a new character, Spiderman, and set the ultimate goal to reach at least 45 points. We finalized the random map generation feature and implemented a win condition that appears on the screen. Additionally, we addressed several bugs, removed unnecessary code, and added more internal comments.